Fix for hardware library reservations

+++ src/util/hardwareLibrary/hardwareItemValidation.ts
@@ -249,7 +249,8 @@ export const getAllHardwareItems = async (userId?: number): Promise<Object[]> =>
   for (const item of hardwareItems) {
     let remainingItemCount: number = item.totalStock - (item.reservedStock + item.takenStock);
     let reservationForItem = userReservations.find(reservation => reservation.hardwareItem.name === item.name);
-    if (reservationForItem && !isReservationValid(reservationForItem.reservationExpiry)) {
+
+    if (reservationForItem && !reservationForItem.isReserved && !isReservationValid(reservationForItem.reservationExpiry)) {
       remainingItemCount += reservationForItem.reservationQuantity;
       await deleteReservation(reservationForItem.reservationToken);
       reservationForItem = undefined;
